pyproject.toml
[tool.poetry]
name = "cta-bus-tracker"
version = "0.1.0"
description = ""
authors = ["Joe Gambino <joe@gam.bingo>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
